[feat] 통계값 추가

+++ src/main/java/com/example/plugissue/station/controller/dto/StationDto.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class StationDto {
 
-    private Long sId;
+    private Integer sId;
 
     private Double lng;
